+++ db.json
@@ -1,14 +1,6 @@
 {
   "todoList": [
     {
-      "id": 1,
-      "todoListTitle": "225521",
-      "user": "22",
-      "editing": false,
-      "todoListPw": "456",
-      "toggle": true
-    },
-    {
       "todoListTitle": "11",
       "user": "22",
       "editing": false,
@@ -29,7 +21,7 @@
       "user": "222",
       "editing": false,
       "todoListPw": "222",
-      "toggle": true,
+      "toggle": false,
       "id": 6
     },
     {
